Upgrade dbcp to v2.9.0 and refactor imports
