+++ modules/commons/pom.xml
@@ -142,8 +142,8 @@
 
     <dependencies>
         <dependency>
-            <groupId>commons-dbcp</groupId>
-            <artifactId>commons-dbcp</artifactId>
+            <groupId>org.apache.commons</groupId>
+            <artifactId>commons-dbcp2</artifactId>
         </dependency>
         <dependency>
             <groupId>org.eclipse.osgi</groupId>
