+++ modules/core/src/main/java/org/apache/synapse/mediators/db/AbstractDBMediator.java
@@ -21,8 +21,8 @@ package org.apache.synapse.mediators.db;
 
 import org.apache.axiom.om.OMText;
 import org.apache.axiom.om.impl.llom.OMTextImpl;
-import org.apache.commons.dbcp.BasicDataSource;
-import org.apache.commons.dbcp.datasources.PerUserPoolDataSource;
+import org.apache.commons.dbcp2.BasicDataSource;
+import org.apache.commons.dbcp2.datasources.PerUserPoolDataSource;
 import org.apache.commons.logging.LogFactory;
 import org.apache.synapse.ManagedLifecycle;
 import org.apache.synapse.MessageContext;
