+++ modules/transports/pom.xml
@@ -50,8 +50,8 @@
             <artifactId>synapse-commons</artifactId>
             <exclusions>
                 <exclusion>
-                    <groupId>commons-dbcp</groupId>
-                    <artifactId>commons-dbcp</artifactId>
+                    <groupId>org.apache.commons</groupId>
+                    <artifactId>commons-dbcp2</artifactId>
                 </exclusion>
             </exclusions>
         </dependency>
