+++ pom.xml
@@ -999,13 +999,13 @@
             <version>${sandesha2.version}</version>
          </dependency>
          <dependency>
-            <groupId>commons-dbcp</groupId>
-            <artifactId>commons-dbcp</artifactId>
+            <groupId>org.apache.commons</groupId>
+            <artifactId>commons-dbcp2</artifactId>
             <version>${commons.dbcp.version}</version>
          </dependency>
          <dependency>
-            <groupId>commons-dbcp.wso2</groupId>
-            <artifactId>commons-dbcp</artifactId>
+            <groupId>org.wso2.orbit.org.apache.commons</groupId>
+            <artifactId>commons-dbcp2</artifactId>
             <version>${commons.dbcp.wso2.version}</version>
          </dependency>
          <dependency>
@@ -1445,8 +1445,8 @@
       <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
       <!-- Synapse and related components -->
       <synapse.version>${project.version}</synapse.version>
-      <commons.dbcp.version>1.2.2</commons.dbcp.version>
-      <commons.dbcp.wso2.version>1.4.0.wso2v1</commons.dbcp.wso2.version>
+      <commons.dbcp.version>2.9.0</commons.dbcp.version>
+      <commons.dbcp.wso2.version>2.9.0.wso2v1</commons.dbcp.wso2.version>
       <commons.pool.version>1.3</commons.pool.version>
       <commons.pool.wso2.version>1.5.0.wso2v1</commons.pool.wso2.version>
       <commons.vfs.version>2.2.0-wso2v13</commons.vfs.version>
